requirements.txt
ultralytics
flask
waitress
pillow
streamlit 
opencv-python-headless